% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mWaveQTL.R
\name{plot_mWaveQTL}
\alias{plot_mWaveQTL}
\title{Data visualisation for ffw output}
\usage{
plot_mWaveQTL(res, bp, lev_res, fill, dg, BF_lev)
}
\arguments{
\item{res}{Output of mWaveQTL.}

\item{bp}{a vector of the base pairs position of the loci, you can provide only the starting point and the end point of the loci. If missing set as 0, 1.}

\item{lev_res}{the maximum level of resolution needed, has to be less or equal to the request level of resolution in the ffw function.}

\item{fill}{logical, if not provide set as TRUE.}

\item{dg}{numerical, the number of digits display on the x axe. If missing set at 3.}

\item{BF_lev}{Level of Bayes Factor use to overlay regions. IF missing set as 1.}
}
\value{
return a ggplot
}
\description{
Data visualisation of mWaveQTL screening results
}
\details{
The function generate a ggplot from the ffw function output. It represents the Bayes factor for the different levels scales of the wavelets decomposition.
The size and the darkness of the points that represent the Bayes factor are scaled by the value of the Bayes factors.
If a Bayes factor is greater than 1 then the region that represent the Bayes factor is filled up in order to give an orverview of the size and the origin of the genetic signal.
}
\seealso{
\code{\link{mWaveQTL}}
}
